move validation to extra file
